requirement.txt
langchain_community 
pypdf
langchain_text_splitters
langchain-openai
python-dotenv
langchain-qdrant

